ai/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
openai>=1.6.1
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-deepseek
langgraph>=0.0.20
langchain-mcp>=0.1.0
scikit-learn>=1.3.2
numpy>=1.25.2
graphrag>=0.5
aiohttp>=3.9.0
requests>=2.31.0
python-multipart>=0.0.6
loguru>=0.7.2
datasets>=2.14.0
fastmcp>=2.3.0
mcp>=1.8.0
sse-starlette>=1.6.0
asyncio>=3.4.3
